move print function to separate file
